Move ubatch-related utils to ubatch_utils
